refactor(studio): extract reconciliation runner
