+++ lib/core/llm/studio_ledger_service.dart
@@ -1,7 +1,6 @@
 import 'dart:async';
 
 import 'package:dio/dio.dart';
-import 'package:flutter/foundation.dart';
 
 import '../db/repositories/character_knowledge_fact_repo.dart';
 import '../db/repositories/character_repo.dart';
@@ -16,8 +15,6 @@ import '../db/repositories/tracker_repo.dart';
 import '../db/repositories/tracker_snapshot_repo.dart';
 import '../models/pipeline_settings.dart';
 import '../models/studio_config.dart';
-import '../models/studio_ledger_export.dart';
-import '../utils/id_generator.dart';
 import '../services/card_rewriter/effective_canon_context_loader.dart';
 import 'aux_llm_client.dart';
 import 'ledger/ledger_canon_authority.dart';
@@ -26,16 +23,15 @@ import 'ledger/ledger_op_applier.dart';
 import 'ledger/ledger_output_recovery.dart';
 import 'ledger/ledger_prompt_factory.dart';
 import 'ledger/ledger_reconciliation_committer.dart';
+import 'ledger/ledger_reconciliation_runner.dart';
 import 'ledger/ledger_replacement_basis_resolver.dart';
 import 'ledger/ledger_run_diagnostics.dart';
 import 'ledger/ledger_run_result.dart';
 import 'ledger/ledger_turn_committer.dart';
 import 'ledger/ledger_turn_runner.dart';
-import 'knowledge_cleanup_parser.dart';
 import 'macro_engine.dart';
 import 'studio_ledger_export_parser.dart';
 import 'studio_ledger_reconciliation.dart';
-import 'transport/llm_capture_context.dart';
 
 export 'ledger/ledger_op_applier.dart';
 export 'ledger/ledger_run_result.dart';
@@ -83,7 +79,6 @@ export 'ledger/ledger_run_result.dart';
 ///
 /// Constructor-injected deps (no `Ref` — all repos/client are injected).
 class StudioLedgerService {
-  static const _reconciliationLeaseTtlSeconds = 60 * 60;
   final AuxLlmClient _llm;
   final TrackerRepo _trackerRepo;
   final MemoryBookRepo _bookRepo;
@@ -92,7 +87,6 @@ class StudioLedgerService {
   final LedgerReconciliationCheckpointRepo _reconciliationCheckpointRepo;
   final LedgerReconciliationRunRepo _reconciliationRunRepo;
   final LedgerCanonAuthority _canonAuthority;
-  final StudioLedgerExportParser _parser;
   final LedgerInFlightRegistry _inFlightRegistry;
   final LedgerOutputRecovery _outputRecovery;
   final LedgerRunDiagnostics _runDiagnostics;
@@ -100,6 +94,7 @@ class StudioLedgerService {
   final ReconciliationReplacementRepo _replacementRepo;
   late final LedgerReplacementBasisResolver _replacementBasisResolver;
   late final LedgerReconciliationCommitter _reconciliationCommitter;
+  late final LedgerReconciliationRunner _reconciliationRunner;
   late final LedgerTurnRunner _turnRunner;
 
   StudioLedgerService({
@@ -123,9 +118,9 @@ class StudioLedgerService {
     LedgerCanonAuthority? canonAuthority,
     LedgerReplacementBasisResolver? replacementBasisResolver,
     LedgerReconciliationCommitter? reconciliationCommitter,
+    LedgerReconciliationRunner? reconciliationRunner,
     LedgerTurnRunner? turnRunner,
-  }) : _parser = const StudioLedgerExportParser(),
-       _inFlightRegistry = inFlightRegistry ?? const LedgerInFlightRegistry(),
+  }) : _inFlightRegistry = inFlightRegistry ?? const LedgerInFlightRegistry(),
        _outputRecovery = outputRecovery ?? const LedgerOutputRecovery(),
        _runDiagnostics =
            runDiagnostics ??
@@ -168,6 +163,19 @@ class StudioLedgerService {
           replacementBasisResolver: _replacementBasisResolver,
           opApplier: const LedgerOpApplier(),
         );
+    _reconciliationRunner =
+        reconciliationRunner ??
+        LedgerReconciliationRunner(
+          llm: _llm,
+          snapshotRepo: _snapshotRepo,
+          reconciliationLeaseRepo: _reconciliationLeaseRepo,
+          canonAuthority: _canonAuthority,
+          outputRecovery: _outputRecovery,
+          committer: _reconciliationCommitter,
+          replacementBasisResolver: _replacementBasisResolver,
+          runDiagnostics: _runDiagnostics,
+          parser: const StudioLedgerExportParser(),
+        );
     _turnRunner =
         turnRunner ??
         LedgerTurnRunner(
@@ -198,18 +206,21 @@ class StudioLedgerService {
     CancelToken? cancelToken,
     String? operationIdentity,
   }) {
+    final request = LedgerReconciliationRequest(
+      sessionId: sessionId,
+      settings: settings,
+      config: config,
+      plan: plan,
+      ledgerBlocks: ledgerBlocks,
+      macroCtx: macroCtx,
+      isStillCurrent: isStillCurrent,
+      cancelToken: cancelToken,
+      purpose: operationIdentity?.startsWith('manual:') == true
+          ? 'manual'
+          : 'normal',
+    );
     if (operationIdentity == null) {
-      return _reconcileUnshared(
-        sessionId: sessionId,
-        settings: settings,
-        config: config,
-        plan: plan,
-        ledgerBlocks: ledgerBlocks,
-        macroCtx: macroCtx,
-        isStillCurrent: isStillCurrent,
-        cancelToken: cancelToken,
-        purpose: 'normal',
-      );
+      return _reconciliationRunner.reconcile(request);
     }
     final end = plan.endMessage;
     final key = _inFlightRegistry.reconciliationKey(
@@ -226,17 +237,7 @@ class StudioLedgerService {
     );
     return _inFlightRegistry.join(
       key,
-      () => _reconcileUnshared(
-        sessionId: sessionId,
-        settings: settings,
-        config: config,
-        plan: plan,
-        ledgerBlocks: ledgerBlocks,
-        macroCtx: macroCtx,
-        isStillCurrent: isStillCurrent,
-        cancelToken: cancelToken,
-        purpose: operationIdentity.startsWith('manual:') ? 'manual' : 'normal',
-      ),
+      () => _reconciliationRunner.reconcile(request),
     );
   }
 
@@ -254,6 +255,16 @@ class StudioLedgerService {
     CancelToken? cancelToken,
     String? operationIdentity,
   }) {
+    final request = LedgerReconciliationReplacementRequest(
+      sessionId: sessionId,
+      expectedRunId: expectedRunId,
+      settings: settings,
+      config: config,
+      ledgerBlocks: ledgerBlocks,
+      macroCtx: macroCtx,
+      isStillCurrent: isStillCurrent,
+      cancelToken: cancelToken,
+    );
     final key = _inFlightRegistry.replacementKey(
       operationIdentity: operationIdentity,
       sessionId: sessionId,
@@ -265,605 +276,9 @@ class StudioLedgerService {
     );
     return _inFlightRegistry.join(
       key,
-      () => _replaceLatestReconciliationUnshared(
-        sessionId: sessionId,
-        expectedRunId: expectedRunId,
-        settings: settings,
-        config: config,
-        ledgerBlocks: ledgerBlocks,
-        macroCtx: macroCtx,
-        isStillCurrent: isStillCurrent,
-        cancelToken: cancelToken,
-      ),
+      () => _reconciliationRunner.replaceLatest(request),
     );
   }
-
-  Future<LedgerRunResult> _replaceLatestReconciliationUnshared({
-    required String sessionId,
-    required String expectedRunId,
-    required PipelineSettings settings,
-    required AuxApiConfig config,
-    required List<StudioPresetBlock> ledgerBlocks,
-    required MacroContext? macroCtx,
-    required FutureOr<bool> Function()? isStillCurrent,
-    required CancelToken? cancelToken,
-  }) async {
-    final ownerId = 'ledger-reconciliation-${generateId()}';
-    if (!await _reconciliationLeaseRepo.acquire(
-      sessionId: sessionId,
-      ownerId: ownerId,
-      purpose: 'replacement',
-      ttlSeconds: _reconciliationLeaseTtlSeconds,
-    )) {
-      return const LedgerRunResult(
-        status: 'skipped',
-        error: 'Another reconciliation is already running for this session',
-      );
-    }
-    final token = cancelToken ?? CancelToken();
-    try {
-      final basis = await _replacementBasisResolver.prepare(
-        sessionId: sessionId,
-        expectedRunId: expectedRunId,
-        token: token,
-        isStillCurrent: isStillCurrent,
-      );
-      if (basis is LedgerReplacementBasisFailure) {
-        return LedgerRunResult(status: 'error', error: basis.reason);
-      }
-      final ready = basis as LedgerReplacementBasisReady;
-      final trace = LedgerRunTrace(
-        sessionId: sessionId,
-        kind: LedgerDebugRunKind.reconciliation,
-        messageId: ready.plan.endMessage.id,
-        swipeId: ready.plan.endMessage.swipeId,
-        agentSwipeId: ready.plan.endMessage.agentSwipeId,
-      );
-      final result = await _reconcileTraced(
-        trace: trace,
-        sessionId: sessionId,
-        settings: settings,
-        config: config,
-        plan: ready.plan,
-        ledgerBlocks: ledgerBlocks,
-        macroCtx: macroCtx,
-        isStillCurrent: isStillCurrent,
-        cancelToken: token,
-        canonOverride: ready.beforeCanon,
-        replacement: ready,
-        leaseOwnerId: ownerId,
-      );
-      if (await _reconciliationLeaseRepo.ownsLiveLeaseInTransaction(
-        sessionId: sessionId,
-        ownerId: ownerId,
-      )) {
-        await _runDiagnostics.recordDebugRun(trace, result);
-      }
-      return result;
-    } finally {
-      await _reconciliationLeaseRepo.release(
-        sessionId: sessionId,
-        ownerId: ownerId,
-      );
-    }
-  }
-
-  Future<LedgerRunResult> _reconcileUnshared({
-    required String sessionId,
-    required PipelineSettings settings,
-    required AuxApiConfig config,
-    required LedgerReconciliationPlan plan,
-    List<StudioPresetBlock> ledgerBlocks = const [],
-    MacroContext? macroCtx,
-    FutureOr<bool> Function()? isStillCurrent,
-    CancelToken? cancelToken,
-    required String purpose,
-  }) async {
-    final ownerId = 'ledger-reconciliation-${generateId()}';
-    if (!await _reconciliationLeaseRepo.acquire(
-      sessionId: sessionId,
-      ownerId: ownerId,
-      purpose: purpose,
-      ttlSeconds: _reconciliationLeaseTtlSeconds,
-    )) {
-      return const LedgerRunResult(
-        status: 'skipped',
-        error: 'Another reconciliation is already running for this session',
-      );
-    }
-    final trace = LedgerRunTrace(
-      sessionId: sessionId,
-      kind: LedgerDebugRunKind.reconciliation,
-      messageId: plan.endMessage.id,
-      swipeId: plan.endMessage.swipeId,
-      agentSwipeId: plan.endMessage.agentSwipeId,
-    );
-    try {
-      final result = await _reconcileTraced(
-        trace: trace,
-        sessionId: sessionId,
-        settings: settings,
-        config: config,
-        plan: plan,
-        ledgerBlocks: ledgerBlocks,
-        macroCtx: macroCtx,
-        isStillCurrent: isStillCurrent,
-        cancelToken: cancelToken,
-        leaseOwnerId: ownerId,
-      );
-      if (await _reconciliationLeaseRepo.ownsLiveLeaseInTransaction(
-        sessionId: sessionId,
-        ownerId: ownerId,
-      )) {
-        await _runDiagnostics.recordDebugRun(trace, result);
-      }
-      return result;
-    } finally {
-      await _reconciliationLeaseRepo.release(
-        sessionId: sessionId,
-        ownerId: ownerId,
-      );
-    }
-  }
-
-  Future<LedgerRunResult> _reconcileTraced({
-    required LedgerRunTrace trace,
-    required String sessionId,
-    required PipelineSettings settings,
-    required AuxApiConfig config,
-    required LedgerReconciliationPlan plan,
-    List<StudioPresetBlock> ledgerBlocks = const [],
-    MacroContext? macroCtx,
-    FutureOr<bool> Function()? isStillCurrent,
-    CancelToken? cancelToken,
-    LedgerCanonContext? canonOverride,
-    LedgerReplacementBasisReady? replacement,
-    required String leaseOwnerId,
-  }) async {
-    final token = cancelToken ?? CancelToken();
-    if (token.isCancelled ||
-        await _canonAuthority.passesCurrentnessGuard(isStillCurrent) == false) {
-      return LedgerRunResult.aborted;
-    }
-    final sw = Stopwatch()..start();
-    try {
-      final canon = canonOverride ?? await _canonAuthority.load(sessionId);
-      await _throwIfReconciliationAborted(token, isStillCurrent);
-      final endpointSnapshot = await _snapshotRepo.getByAnchor(
-        sessionId: sessionId,
-        messageId: plan.endMessage.id,
-        swipeId: plan.endMessage.swipeId,
-        agentSwipeId: plan.endMessage.agentSwipeId,
-      );
-      await _throwIfReconciliationAborted(token, isStillCurrent);
-      if (endpointSnapshot == null || !endpointSnapshot.committed) {
-        return LedgerRunResult(
-          status: 'skipped',
-          error: 'review endpoint snapshot is not committed',
-          elapsedMs: sw.elapsedMilliseconds,
-        );
-      }
-
-      final promptTrackers = _canonAuthority.projectPromptTrackers(
-        canon.context,
-      );
-      final reviewMessageIds = plan.messageIds.toSet();
-      final reviewableFacts = canon.context.resolution.activeFacts;
-      final promptFacts = reviewableFacts
-          .where((fact) => reviewMessageIds.contains(fact.sourceMessageId))
-          .toList();
-      final duplicateRetractions = exactDuplicateKnowledgeRetractions(
-        reviewableFacts,
-      );
-      final staleAnchorRetractions = staleKnowledgeAnchorRetractions(
-        reviewableFacts,
-        plan.messages,
-      );
-      await _throwIfReconciliationAborted(token, isStillCurrent);
-      final promptBlock = ledgerBlocks
-          .where(
-            (block) =>
-                block.id == ledgerReconciliationPromptBlockId &&
-                block.enabled &&
-                block.injectionPoint == 'ledger' &&
-                block.content.trim().isNotEmpty,
-          )
-          .firstOrNull;
-      final systemPrompt = promptBlock == null
-          ? fallbackLedgerReconciliationPrompt
-          : macroCtx == null
-          ? promptBlock.content
-          : replaceMacros(promptBlock.content, macroCtx).text;
-      const reconciliationPrompt = StudioLedgerReconciliationPrompt();
-      final reviewText = plan.messages
-          .map((message) => message.content)
-          .join('\n');
-      final offeredFacts = reconciliationPrompt.relevantKnowledgeFacts(
-        promptFacts,
-        reviewText,
-      );
-      final prompt = reconciliationPrompt.build(
-        systemPrompt: systemPrompt,
-        plan: plan,
-        trackers: promptTrackers,
-        knowledgeFacts: offeredFacts,
-        character: canon.source,
-      );
-      await _throwIfReconciliationAborted(token, isStillCurrent);
-      if (!await _renewReconciliationLease(sessionId, leaseOwnerId)) {
-        return LedgerRunResult.aborted;
-      }
-      final timeoutMs = _llm.resolveLedgerTimeout(settings);
-      final outcome = await _llm.callOnceWithLog(
-        config: config,
-        prompt: prompt,
-        maxTokens: settings.ledger.studioLedgerMaxTokens > 0
-            ? settings.ledger.studioLedgerMaxTokens
-            : 15000,
-        temperature: settings.ledger.studioLedgerTemperature >= 0
-            ? settings.ledger.studioLedgerTemperature
-            : 0.2,
-        timeoutMs: timeoutMs,
-        cancelToken: token,
-        omitReasoning: true,
-        captureContext: LlmCaptureContext(
-          stage: 'ledger.reconciliation',
-          sessionId: sessionId,
-          messageId: plan.endMessage.id,
-          pipelineRunId: plan.rangeHash,
-          logicalCallId: plan.rangeHash,
-          relatedArtifactId: plan.rangeHash,
-        ),
-      );
-      if (token.isCancelled ||
-          await _canonAuthority.passesCurrentnessGuard(isStillCurrent) ==
-              false) {
-        return LedgerRunResult.aborted;
-      }
-      if (!await _replacementBasisResolver.isReconciliationBasisCurrent(
-        sessionId: sessionId,
-        canon: canon,
-        replacement: replacement,
-      )) {
-        return LedgerRunResult.aborted;
-      }
-      if (!outcome.isOk || outcome.text == null || outcome.text!.isEmpty) {
-        final attempt = outcome.attempts.lastOrNull;
-        return LedgerRunResult(
-          status: 'error',
-          error: 'Reconciliation LLM call failed: ${attempt?.status}',
-          elapsedMs: sw.elapsedMilliseconds,
-          attempts: outcome.attempts,
-          model: config.model,
-        );
-      }
-
-      var responseText = outcome.text!;
-      var cleanupResponseText = responseText;
-      var parsed = _parser.parse(
-        responseText,
-        focalUserName: macroCtx?.userName ?? '',
-      );
-      await _runDiagnostics.recordLedgerParserVerdict(outcome, parsed);
-      trace.recordFirstResponse(
-        model: config.model,
-        responseText: responseText,
-        parsed: parsed,
-      );
-      final originalFailure = parsed.failure;
-      final originalVisibleLedger = parsed.visibleLedger;
-      var attempts = outcome.attempts;
-      var repairAttempted = false;
-      var totalPromptChars = prompt.length;
-      var totalResponseChars = responseText.length;
-      const cleanupParser = KnowledgeCleanupParser();
-      final needsExportRepair = parsed.failure.isRepairable;
-      final needsCleanupRepair =
-          (parsed.hasExport ||
-              parsed.failure == LedgerParseFailure.emptyExport) &&
-          !cleanupParser.hasValidBlock(responseText);
-      final needsRepair = needsExportRepair || needsCleanupRepair;
-      if (needsRepair) {
-        await _throwIfReconciliationAborted(token, isStillCurrent);
-        if (!await _replacementBasisResolver.isReconciliationBasisCurrent(
-          sessionId: sessionId,
-          canon: canon,
-          replacement: replacement,
-        )) {
-          return LedgerRunResult.aborted;
-        }
-        repairAttempted = true;
-        trace.recordRepairRequested(exportRepair: needsExportRepair);
-        if (_outputRecovery.isOversizedRepairInput(responseText)) {
-          return LedgerRunResult(
-            status: 'error',
-            visibleLedger: originalVisibleLedger,
-            error: 'Ledger output is too large for safe deterministic repair',
-            elapsedMs: sw.elapsedMilliseconds,
-            attempts: attempts,
-            model: config.model,
-            effectiveTimeoutMs: timeoutMs,
-            promptChars: totalPromptChars,
-            responseChars: totalResponseChars,
-          );
-        }
-        final repairPrompt = needsExportRepair
-            ? _outputRecovery.buildRepairPrompt(
-                responseText,
-                reconciliation: true,
-              )
-            : _outputRecovery.buildCleanupRepairPrompt(responseText);
-        totalPromptChars += repairPrompt.length;
-        if (!await _renewReconciliationLease(sessionId, leaseOwnerId)) {
-          return LedgerRunResult.aborted;
-        }
-        final repair = await _llm.callOnceWithLog(
-          config: config,
-          prompt: repairPrompt,
-          maxTokens: settings.ledger.studioLedgerMaxTokens > 0
-              ? settings.ledger.studioLedgerMaxTokens
-              : 15000,
-          temperature: 0,
-          timeoutMs: timeoutMs,
-          cancelToken: token,
-          omitReasoning: true,
-          captureContext: LlmCaptureContext(
-            stage: 'ledger.reconciliation_repair',
-            sessionId: sessionId,
-            messageId: plan.endMessage.id,
-            pipelineRunId: plan.rangeHash,
-            logicalCallId: '${plan.rangeHash}:repair',
-            relatedArtifactId: plan.rangeHash,
-          ),
-        );
-        attempts = _outputRecovery.combineAttempts(attempts, repair.attempts);
-        totalResponseChars += repair.text?.length ?? 0;
-        await _throwIfReconciliationAborted(token, isStillCurrent);
-        if (!await _replacementBasisResolver.isReconciliationBasisCurrent(
-          sessionId: sessionId,
-          canon: canon,
-          replacement: replacement,
-        )) {
-          return LedgerRunResult.aborted;
-        }
-        if (!repair.isOk || repair.text == null || repair.text!.isEmpty) {
-          return LedgerRunResult(
-            status: 'error',
-            error: 'Ledger export repair call failed',
-            elapsedMs: sw.elapsedMilliseconds,
-            attempts: attempts,
-            model: config.model,
-            repairAttempted: true,
-            effectiveTimeoutMs: timeoutMs,
-            promptChars: totalPromptChars,
-            responseChars: totalResponseChars,
-          );
-        }
-        if (needsExportRepair) {
-          final repairedText = repair.text!;
-          final repaired = _parser.parse(
-            repairedText,
-            focalUserName: macroCtx?.userName ?? '',
-          );
-          await _runDiagnostics.recordLedgerParserVerdict(repair, repaired);
-          trace.recordRepairResponse(
-            responseText: repairedText,
-            parsed: repaired,
-          );
-          if (repaired.export != null &&
-              !_outputRecovery.repairPreservesStructuredEvidence(
-                responseText,
-                repaired.export!,
-              )) {
-            return LedgerRunResult(
-              status: 'error',
-              visibleLedger: originalVisibleLedger,
-              error:
-                  'Ledger repair introduced data absent from the original output',
-              elapsedMs: sw.elapsedMilliseconds,
-              attempts: attempts,
-              model: config.model,
-              repairAttempted: true,
-              effectiveTimeoutMs: timeoutMs,
-              promptChars: totalPromptChars,
-              responseChars: totalResponseChars,
-            );
-          }
-          responseText = repairedText;
-          cleanupResponseText = repairedText;
-          parsed = repaired;
-          final repairedCleanup = cleanupParser.parse(
-            output: cleanupResponseText,
-            offeredFacts: offeredFacts,
-            reviewText: reviewText,
-          );
-          if (!_outputRecovery.cleanupRepairPreservesLiteralEvidence(
-            outcome.text!,
-            repairedCleanup,
-          )) {
-            return LedgerRunResult(
-              status: 'error',
-              visibleLedger: originalVisibleLedger,
-              error:
-                  'Ledger cleanup repair introduced data absent from the original output',
-              elapsedMs: sw.elapsedMilliseconds,
-              attempts: attempts,
-              model: config.model,
-              repairAttempted: true,
-              effectiveTimeoutMs: timeoutMs,
-              promptChars: totalPromptChars,
-              responseChars: totalResponseChars,
-            );
-          }
-        } else {
-          // The export was already valid. A cleanup-only repair must never
-          // replace or reinterpret the authoritative tracker export.
-          cleanupResponseText = repair.text!;
-          trace.recordRepairResponse(responseText: cleanupResponseText);
-          final repairedCleanup = cleanupParser.parse(
-            output: cleanupResponseText,
-            offeredFacts: offeredFacts,
-            reviewText: reviewText,
-          );
-          if (!_outputRecovery.cleanupRepairPreservesLiteralEvidence(
-            responseText,
-            repairedCleanup,
-          )) {
-            return LedgerRunResult(
-              status: 'error',
-              visibleLedger: originalVisibleLedger,
-              error:
-                  'Ledger cleanup repair introduced data absent from the original output',
-              elapsedMs: sw.elapsedMilliseconds,
-              attempts: attempts,
-              model: config.model,
-              repairAttempted: true,
-              effectiveTimeoutMs: timeoutMs,
-              promptChars: totalPromptChars,
-              responseChars: totalResponseChars,
-            );
-          }
-        }
-      }
-      final isEmptyExport =
-          parsed.rejectionReason == 'empty export (no ops or knowledge facts)';
-      if (repairAttempted &&
-          originalFailure == LedgerParseFailure.missingExport &&
-          parsed.failure == LedgerParseFailure.emptyExport) {
-        return LedgerRunResult(
-          status: 'error',
-          visibleLedger: originalVisibleLedger,
-          error:
-              'Repair produced an empty export without explicit empty intent',
-          elapsedMs: sw.elapsedMilliseconds,
-          attempts: attempts,
-          model: config.model,
-          repairAttempted: true,
-          effectiveTimeoutMs: timeoutMs,
-          promptChars: totalPromptChars,
-          responseChars: totalResponseChars,
-        );
-      }
-      if (!parsed.hasExport && !isEmptyExport) {
-        return LedgerRunResult(
-          status: 'error',
-          visibleLedger: originalVisibleLedger.isNotEmpty
-              ? originalVisibleLedger
-              : parsed.visibleLedger,
-          error: parsed.rejectionReason,
-          elapsedMs: sw.elapsedMilliseconds,
-          attempts: attempts,
-          model: config.model,
-          repairAttempted: repairAttempted,
-          effectiveTimeoutMs: timeoutMs,
-          promptChars: totalPromptChars,
-          responseChars: totalResponseChars,
-        );
-      }
-      final export = parsed.export ?? const StudioLedgerExport();
-      if (export.knowledgeFacts.isNotEmpty) {
-        return LedgerRunResult(
-          status: 'error',
-          error: 'Reconciliation must not emit knowledgeFacts',
-          elapsedMs: sw.elapsedMilliseconds,
-          attempts: attempts,
-          model: config.model,
-          repairAttempted: repairAttempted,
-          effectiveTimeoutMs: timeoutMs,
-          promptChars: totalPromptChars,
-          responseChars: totalResponseChars,
-        );
-      }
-      if (token.isCancelled ||
-          await _canonAuthority.passesCurrentnessGuard(isStillCurrent) ==
-              false) {
-        return LedgerRunResult.aborted;
-      }
-      if (!cleanupParser.hasValidBlock(cleanupResponseText)) {
-        return LedgerRunResult(
-          status: 'error',
-          error: 'Reconciliation returned no valid knowledge cleanup block',
-          elapsedMs: sw.elapsedMilliseconds,
-          attempts: attempts,
-          model: config.model,
-          repairAttempted: repairAttempted,
-        );
-      }
-      final cleanupOps =
-          cleanupParser.parse(
-              output: cleanupResponseText,
-              offeredFacts: offeredFacts,
-              reviewText: reviewText,
-            )
-            ..addAll(duplicateRetractions)
-            ..addAll(staleAnchorRetractions);
-      final allowedCleanupFactIds = {
-        ...offeredFacts.map((fact) => fact.id),
-        ...duplicateRetractions.map((op) => op.factId),
-        ...staleAnchorRetractions.map((op) => op.factId),
-      };
-      final commit = await _reconciliationCommitter.commit(
-        LedgerReconciliationCommitRequest(
-          sessionId: sessionId,
-          leaseOwnerId: leaseOwnerId,
-          plan: plan,
-          canon: canon,
-          promptTrackers: promptTrackers,
-          export: export,
-          cleanupOps: cleanupOps,
-          allowedCleanupFactIds: allowedCleanupFactIds,
-          token: token,
-          isStillCurrent: isStillCurrent,
-          replacement: replacement,
-        ),
-      );
-      return LedgerRunResult(
-        status: 'ok',
-        visibleLedger: originalVisibleLedger.isNotEmpty
-            ? originalVisibleLedger
-            : parsed.visibleLedger,
-        opsApplied: commit.replayed ? 0 : commit.opsApplied,
-        elapsedMs: sw.elapsedMilliseconds,
-        attempts: attempts,
-        model: config.model,
-        repairAttempted: repairAttempted,
-        effectiveTimeoutMs: timeoutMs,
-        promptChars: totalPromptChars,
-        responseChars: totalResponseChars,
-      );
-    } on LedgerCommitStale {
-      return LedgerRunResult.aborted;
-    } on _LedgerReconciliationAborted {
-      return LedgerRunResult.aborted;
-    } catch (e) {
-      if (token.isCancelled || (e is DioException && CancelToken.isCancel(e))) {
-        return LedgerRunResult.aborted;
-      }
-      debugPrint('[StudioLedger] reconciliation failed: $e');
-      return LedgerRunResult(
-        status: 'error',
-        error: '$e',
-        elapsedMs: sw.elapsedMilliseconds,
-      );
-    }
-  }
-
-  Future<void> _throwIfReconciliationAborted(
-    CancelToken token,
-    FutureOr<bool> Function()? isStillCurrent,
-  ) async {
-    if (token.isCancelled ||
-        await _canonAuthority.passesCurrentnessGuard(isStillCurrent) == false) {
-      throw const _LedgerReconciliationAborted();
-    }
-  }
-
-  Future<bool> _renewReconciliationLease(String sessionId, String ownerId) =>
-      _reconciliationLeaseRepo.renew(
-        sessionId: sessionId,
-        ownerId: ownerId,
-        ttlSeconds: _reconciliationLeaseTtlSeconds,
-      );
 
   /// Run the Studio Ledger for [sessionId] on [finalAssistantText].
   ///
@@ -930,8 +345,4 @@ class StudioLedgerService {
     );
     return _inFlightRegistry.join(key, () => _turnRunner.run(request));
   }
-}
-
-class _LedgerReconciliationAborted implements Exception {
-  const _LedgerReconciliationAborted();
 }
